moved svgs and css to assets folder
