+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
+import reactLogo from './assets/img/react.svg'
 import viteLogo from '/vite.svg'
-import './App.css'
+import './assets/css/App.css'
 
 function App() {
   const [count, setCount] = useState(0)
