+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
-import './index.css'
+import './assets/css/index.css'
 
 async function enableMocking() {
   if (process.env.NODE_ENV !== 'development') {
